refactor: move all importer files from lib to src. Make all files PSR-4 compliant. Format all files in accordance with AivecWP-5 ruleset
