+++ plugin.php
@@ -22,10 +22,6 @@ if (file_exists(__DIR__ . '/vendor/autoload.php')) {
     require __DIR__ . '/vendor/autoload.php';
 }
 
-global $wp_parser;
-$wp_parser = new WP_Parser\Plugin();
-$wp_parser->on_load();
-
 Aivec\Plugins\DocParser\Master::init();
 
 add_filter('wp_parser_exclude_directories', function () {
@@ -39,7 +35,7 @@ add_filter('wp_parser_exclude_directories_strict', function () {
 register_activation_hook(__FILE__, ['P2P_Storage', 'init']);
 register_activation_hook(__FILE__, ['P2P_Storage', 'install']);
 register_activation_hook(__FILE__, function () {
-    (new WP_Parser\Plugin())->register_post_types();
+    (new Aivec\Plugins\DocParser\Registrations())->registerPostTypes();
     flush_rewrite_rules();
 });
